Update with a reusable gdbus_library() template
