+++ src/platform/BUILD.gn
@@ -18,29 +18,6 @@ import("//build/config/linux/pkg_config.gni")
 
 import("device.gni")
 
-action("gen_gdbus_wrapper") {
-  script = "${chip_root}/scripts/gen_gdbus_wrapper.py"
-
-  source_path = "//src/platform/Linux/dbus"
-  gen_dbus_wpa = "${target_gen_dir}/Linux/dbus/DBusWpa.c"
-  gen_dbus_wpa_interface = "${target_gen_dir}/Linux/dbus/DBusWpaInterface.c"
-  gen_dbus_wpa_network = "${target_gen_dir}/Linux/dbus/DBusWpaNetwork.c"
-  outputs = [ gen_dbus_wpa,  gen_dbus_wpa_interface, gen_dbus_wpa_network]
-
-  outputs_files  = rebase_path("${gen_dbus_wpa}", root_build_dir) + ","
-  outputs_files += rebase_path("${gen_dbus_wpa_interface}", root_build_dir) + ","
-  outputs_files += rebase_path("${gen_dbus_wpa_network}", root_build_dir)
-
-  inputs_files  = rebase_path("${source_path}/wpa.xml", root_build_dir) + ","
-  inputs_files += rebase_path("${source_path}/wpa_interface.xml", root_build_dir) + ","
-  inputs_files += rebase_path("${source_path}/wpa_network.xml", root_build_dir)
-
-  args = [
-    "--output_files=" + outputs_files,
-    "--input_files=" + inputs_files,
-  ]
-}
-
 if (chip_device_platform == "nrf5") {
   import("//build_overrides/nrf5_sdk.gni")
 }
@@ -68,7 +45,6 @@ if (chip_device_platform != "none") {
     if (chip_device_platform_config_include != "") {
       defines += [ "CHIP_DEVICE_PLATFORM_CONFIG_INCLUDE=${chip_device_platform_config_include}" ]
     }
-
 
     if (chip_device_config_firmware_build_date != "") {
       defines += [ "CHIP_DEVICE_CONFIG_FIRWMARE_BUILD_DATE=\"${chip_device_config_firmware_build_date}\"" ]
@@ -114,7 +90,6 @@ if (chip_device_platform != "none") {
     }
     if (chip_enable_wifi) {
       defines += [ "CHIP_DEVICE_CONFIG_ENABLE_WPA=1" ]
-      include_dirs = [ "${chip_root}/src/platform/Linux/dbus" ]
     }
     if (chip_device_platform == "efr32") {
       defines += [
@@ -125,13 +100,6 @@ if (chip_device_platform != "none") {
     } else {
       defines += [ "CHIP_DEVICE_LAYER_TARGET_EFR32=0" ]
     }
-  }
-
-  pkg_config("glib") {
-    packages = [
-    "glib-2.0",
-    "gio-unix-2.0",
-    ]
   }
 
   static_library("platform") {
@@ -249,16 +217,7 @@ if (chip_device_platform != "none") {
       }
 
       if (chip_enable_wifi) {
-        sources += [
-          "Linux/dbus/DBusWpa.h",
-          "${target_gen_dir}/Linux/dbus/DBusWpa.c",
-          "Linux/dbus/DBusWpaInterface.h",
-          "${target_gen_dir}/Linux/dbus/DBusWpaInterface.c",
-          "Linux/dbus/DBusWpaNetwork.h",
-          "${target_gen_dir}/Linux/dbus/DBusWpaNetwork.c",
-        ]
-        deps = [ ":gen_gdbus_wrapper" ]
-        public_configs += [ ":glib" ]
+        public_deps += [ "Linux/dbus" ]
       }
 
       public_deps += [ "${chip_root}/third_party/inipp" ]
