+++ src/platform/Linux/dbus/BUILD.gn
@@ -1,0 +1,29 @@
+# Copyright (c) 2020 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import("//build_overrides/chip.gni")
+
+import("${chip_root}/gn/chip/linux/gdbus_library.gni")
+
+gdbus_library("dbus") {
+  sources = [
+    "DBusWpa.xml",
+    "DBusWpaInterface.xml",
+    "DBusWpaNetwork.xml",
+  ]
+
+  c_namespace = "Wpa"
+
+  dbus_out_dir = "platform/Linux/dbus"
+}
