+++ src/platform/Makefile.am
@@ -246,24 +246,24 @@ BUILT_SOURCES                          += \
     Linux/gen/DBusWpaNetwork.c            \
     $(NULL)
 
-Linux/gen/DBusWpa.h: Linux/dbus/wpa.xml
+Linux/gen/DBusWpa.h: Linux/dbus/DBusWpa.xml
 	gdbus-codegen --header --output $@ --c-namespace Wpa $<
 
-Linux/gen/DBusWpaInterface.h: Linux/dbus/wpa_interface.xml
+Linux/gen/DBusWpaInterface.h: Linux/dbus/DBusWpaInterface.xml
 	gdbus-codegen --header --output $@ --c-namespace Wpa $<
 
-Linux/gen/DBusWpaNetwork.h: Linux/dbus/wpa_network.xml
+Linux/gen/DBusWpaNetwork.h: Linux/dbus/DBusWpaNetwork.xml
 	gdbus-codegen --header --output $@ --c-namespace Wpa $<
 
-Linux/gen/DBusWpa.c: Linux/dbus/wpa.xml
+Linux/gen/DBusWpa.c: Linux/dbus/DBusWpa.xml
 	gdbus-codegen --body --output $@ --c-namespace Wpa $<
 	sed -i 's/config\.h/BuildConfig.h/g' $@
 
-Linux/gen/DBusWpaInterface.c: Linux/dbus/wpa_interface.xml
+Linux/gen/DBusWpaInterface.c: Linux/dbus/DBusWpaInterface.xml
 	gdbus-codegen --body --output $@ --c-namespace Wpa $<
 	sed -i 's/config\.h/BuildConfig.h/g' $@
 
-Linux/gen/DBusWpaNetwork.c: Linux/dbus/wpa_network.xml
+Linux/gen/DBusWpaNetwork.c: Linux/dbus/DBusWpaNetwork.xml
 	gdbus-codegen --body --output $@ --c-namespace Wpa $<
 	sed -i 's/config\.h/BuildConfig.h/g' $@
